wav: Add function for seeking to chunk
As it turns out it's possible that the wave file got some XML embedded and if this is the case our seek_to_pattern function just searches for the first data it can find, leading to not playing the song because data size is read wrong.
The new function seek_to_chunk assumes that every chunk uses 4 bytes for chunk id and next 4 bytes for the size. This way it will jump from chunk to chunk so it can check if it's data or not. Definitely needs a rewrite.